Disambiguation & doc re-org for Classic
With the release of Classic 6.x it no longer makes sense to use "5x" and
other similar constructs as a general way to refer to the Classic
broker. These ambiguous references should be replaced with "Classic,"
etc.

Furthermore, the Classic documentation should be relocated somewhere
under the components/classic directory. It was originally left in the
root directory to avoid 404s if moved.

This commit resolves these issues via the following changes:

 - Reorganize Classic documentation under the
   components/classic/documentation directory.
 - Add .htaccess entries to avoid 404s.
 - Disambiguate "ActiveMQ Classic" from "ActiveMQ".
 - Fix some formatting, broken links, etc.
 - Rename release files for consistency.
 - Rename formatting elements from "activemq5" to "classic".